+++ R/score_hess.R
@@ -8,6 +8,14 @@
 #' @return A function with signature \code{function(df, par, ...)} returning
 #'   the gradient vector.
 #'
+#' @examples
+#' model <- dfr_series_md(components = list(
+#'   dfr_exponential(0.1), dfr_exponential(0.2)
+#' ))
+#' set.seed(1)
+#' df <- rdata(model)(theta = c(0.1, 0.2), n = 50, tau = 10, p = 0.3)
+#' s_fn <- score(model)
+#' s_fn(df, par = c(0.1, 0.2))
 #' @importFrom likelihood.model score
 #' @importFrom numDeriv grad
 #' @method score dfr_series_md
@@ -30,6 +38,14 @@ score.dfr_series_md <- function(model, ...) {
 #' @return A function with signature \code{function(df, par, ...)} returning
 #'   the Hessian matrix.
 #'
+#' @examples
+#' model <- dfr_series_md(components = list(
+#'   dfr_exponential(0.1), dfr_exponential(0.2)
+#' ))
+#' set.seed(1)
+#' df <- rdata(model)(theta = c(0.1, 0.2), n = 50, tau = 10, p = 0.3)
+#' H_fn <- hess_loglik(model)
+#' H_fn(df, par = c(0.1, 0.2))
 #' @importFrom likelihood.model hess_loglik
 #' @importFrom numDeriv hessian
 #' @method hess_loglik dfr_series_md